+++ bot/menu/__init__.py
@@ -1,11 +1,11 @@
 """
 /bot/menu/__init__.py
-Server Monitoring System v8.56.47
+Server Monitoring System v8.56.49
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Menu exports
 Система мониторинга серверов
-Версия: 8.56.47
+Версия: 8.56.49
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Экспорт функций меню